# api/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Fleet Management Platform API
  version: 1.0.0-locked
  description: >-
    Locked API contract for the Enterprise Fleet Management Platform.
    This document is generated from docs/architecture/00-locked-decisions.md and
    the PostgreSQL schema in db/schema. The frontend team binds the externally
    supplied UI designs to these endpoints (decision A3.6). No UI prescription
    is contained here.

    Conventions (decisions D1, D4, D7, C5.1):
      * Base path /api/v1. JSON only.
      * All timestamps are RFC 3339 in UTC (stored as timestamptz, rendered EAT client-side per A2.3).
      * Cursor pagination via `cursor` / `limit` query params; cursors are opaque.
      * Errors use RFC 7807 problem+json with an `error_code` extension member.
      * Every state-changing request MUST carry `Idempotency-Key: <uuid>` (C5.1).
  contact:
    name: Fleet Platform Architecture
servers:
  - url: https://api.fleet.internal/api/v1
    description: Production (AWS af-south-1, decision A1.10)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
  parameters:
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque cursor returned by a previous page.
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem+json.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: https://docs.fleet.internal/problems/clockout_pending }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uuid }
        error_code:
          type: string
          description: Stable machine-readable code (see Error Codes).
          example: CLOCKOUT_PENDING
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    VehicleCreateRequest:
      type: object
      description: >
        Onboards a vehicle into the CALLER's tenant. license_plate is unique per tenant
        (vehicles_tenant_plate_unique) and a duplicate is rejected with 409. tenant_id is bound to
        the principal, status defaults to AVAILABLE, is_operational to true, and tracker_imei stays
        NULL until the separate A1.1 pairing flow runs.
      required: [license_plate]
      properties:
        license_plate: { type: string, minLength: 1, maxLength: 20 }
        vehicle_class: { type: string, maxLength: 40 }
        make: { type: string, maxLength: 60 }
        model: { type: string, maxLength: 60 }
        year: { type: integer, minimum: 1950, maximum: 2100 }
        ownership_type: { type: string, maxLength: 40 }
        fuel_tank_capacity_litres: { type: number, exclusiveMinimum: 0, maximum: 5000 }
        notes: { type: string, maxLength: 2000 }
    VehicleUpdateRequest:
      type: object
      description: Patches the narrow set of app.vehicles columns the console owns.
      properties:
        status: { type: string, maxLength: 40 }
        is_operational: { type: boolean }
        notes: { type: string, maxLength: 2000, nullable: true }
        non_operational_reason: { type: string, maxLength: 2000, nullable: true }
    VehicleRecord:
      type: object
      description: One app.vehicles row. Numeric columns are normalised from PG text to numbers.
      required: [id, license_plate]
      properties:
        id: { type: string, format: uuid }
        license_plate: { type: string }
        vehicle_class: { type: string, nullable: true }
        status: { type: string, nullable: true }
        is_operational: { type: boolean, nullable: true }
        non_operational_reason: { type: string, nullable: true }
        make: { type: string, nullable: true }
        model: { type: string, nullable: true }
        year: { type: integer, nullable: true }
        ownership_type: { type: string, nullable: true }
        current_odometer_km: { type: number, nullable: true }
        current_odometer_at: { type: string, nullable: true }
        engine_hours: { type: number, nullable: true }
        fuel_tank_capacity_litres: { type: number, nullable: true }
        notes: { type: string, nullable: true }
...
  responses:
    BadRequest:
      description: Malformed request (RFC7807).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Forbidden:
      description: Authenticated but lacking permission.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Resource not found.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Optimistic / state conflict.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
paths:
...
  /vehicles:
    get:
      tags: [Vehicles]
      summary: The caller's fleet (tenant-scoped, cursor-paged).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Page of vehicles.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/VehicleRecord' }
                  next_cursor: { type: string, nullable: true }
                  has_more: { type: boolean }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Vehicles]
      summary: Onboard a vehicle into the caller's tenant (A1.3).
      security:
        - bearerAuth: []
        - idempotencyKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleCreateRequest' }
      responses:
        '201':
          description: Vehicle created.
          content: { application/json: { schema: { $ref: '#/components/schemas/VehicleRecord' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /vehicles/{vehicleId}:
    get:
      tags: [Vehicles]
      summary: One vehicle in the caller's tenant.
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The vehicle.
          content: { application/json: { schema: { $ref: '#/components/schemas/VehicleRecord' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Vehicles]
      summary: Edit the vehicle columns the console owns.
      security:
        - bearerAuth: []
        - idempotencyKey: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleUpdateRequest' }
      responses:
        '200':
          description: The updated vehicle.
          content: { application/json: { schema: { $ref: '#/components/schemas/VehicleRecord' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...